refactor(anta): move AntaDevice and its subclass to anta.device module
